+++ src/lib/core/capability.ts
@@ -4,7 +4,7 @@
 import { GenericClass, GroupVersionKind, modelToGroupVersionKind } from "kubernetes-fluent-client";
 import { pickBy } from "ramda";
 import Log from "../telemetry/logger";
-import { isBuildMode, isDevMode, isWatchMode } from "./module";
+import { isBuildMode, isDevMode, isWatchMode } from "./envChecks";
 import { PeprStore, Storage } from "./storage";
 import { OnSchedule, Schedule } from "./schedule";
 import { Event } from "../enums";
