+++ src/lib/core/module.ts
@@ -4,71 +4,12 @@ import { clone } from "ramda";
 import { Capability } from "./capability";
 import { Controller, ControllerHooks } from "../controller";
 import { ValidateError } from "../errors";
-import { MutateResponse, ValidateResponse, WebhookIgnore } from "../k8s";
-import { CapabilityExport, AdmissionRequest } from "../types";
+import { CapabilityExport } from "../types";
 import { setupWatch } from "../processors/watch-processor";
 import { Log } from "../../lib";
-import { V1PolicyRule as PolicyRule } from "@kubernetes/client-node";
 import { resolveIgnoreNamespaces } from "../assets/webhooks";
-
-/** Custom Labels Type for package.json */
-
-export type CustomLabels = { namespace: Record<string, string> } | Record<string, never>;
-
-/** Configuration that MAY be set a Pepr module's package.json. */
-export type ModuleConfigOptions = {
-  /** The Pepr version this module uses */
-  peprVersion: string;
-  /** The user-defined version of the module */
-  appVersion: string;
-  /** A description of the Pepr module and what it does. */
-  description: string;
-  /** The webhookTimeout */
-  webhookTimeout: number;
-  /** Reject K8s resource AdmissionRequests on error. */
-  onError: string;
-  /** Define the log level for the in-cluster controllers */
-  logLevel: string;
-  /** Propagate env variables to in-cluster controllers */
-  env: Record<string, string>;
-  /** Custom RBAC rules */
-  rbac: PolicyRule[];
-  /** The RBAC mode; if "scoped", generates scoped rules, otherwise uses wildcard rules. */
-  rbacMode: string;
-  /** Custom Labels for Kubernetes Objects */
-  customLabels: CustomLabels;
-};
-
-/** Global configuration for the Pepr runtime. */
-export type ModuleConfig = {
-  /** A unique identifier for this Pepr module. This is automatically generated by Pepr. */
-  uuid: string;
-  /** Configure global exclusions that will never be processed by Pepr. */
-  alwaysIgnore: WebhookIgnore;
-} & Partial<ModuleConfigOptions>;
-
-export type PackageJSON = {
-  description: string;
-  pepr: ModuleConfig;
-};
-
-export type PeprModuleOptions = {
-  deferStart?: boolean;
-
-  /** A user-defined callback to pre-process or intercept a Pepr request from K8s immediately before it is processed */
-  beforeHook?: (req: AdmissionRequest) => void;
-
-  /** A user-defined callback to post-process or intercept a Pepr response just before it is returned to K8s */
-  afterHook?: (res: MutateResponse | ValidateResponse) => void;
-};
-
-// Track if this is a watch mode controller
-export const isWatchMode = (): boolean => process.env.PEPR_WATCH_MODE === "true";
-
-// Track if Pepr is running in build mode
-export const isBuildMode = (): boolean => process.env.PEPR_MODE === "build";
-
-export const isDevMode = (): boolean => process.env.PEPR_MODE === "dev";
+import { isBuildMode, isDevMode, isWatchMode } from "./envChecks";
+import { PackageJSON, PeprModuleOptions, ModuleConfig } from "../types";
 
 export class PeprModule {
   #controller!: Controller;
