+++ src/lib/types.ts
@@ -8,6 +8,7 @@ import { Logger } from "pino";
 import { PeprMutateRequest } from "./mutate-request";
 import { PeprValidateRequest } from "./validate-request";
 import { V1PolicyRule as PolicyRule } from "@kubernetes/client-node";
+import { WebhookIgnore, MutateResponse, ValidateResponse } from "./k8s";
 
 /**
  * Specifically for deploying images with a private registry
@@ -356,4 +357,51 @@ export interface GroupVersionResource {
 // DeepPartial utility type for deep optional properties
 export type DeepPartial<T> = {
   [P in keyof T]?: T[P] extends object ? DeepPartial<T[P]> : T[P];
+}; /** Custom Labels Type for package.json */
+
+export type CustomLabels = { namespace: Record<string, string> } | Record<string, never>;
+/** Configuration that MAY be set a Pepr module's package.json. */
+export type ModuleConfigOptions = {
+  /** The Pepr version this module uses */
+  peprVersion: string;
+  /** The user-defined version of the module */
+  appVersion: string;
+  /** A description of the Pepr module and what it does. */
+  description: string;
+  /** The webhookTimeout */
+  webhookTimeout: number;
+  /** Reject K8s resource AdmissionRequests on error. */
+  onError: string;
+  /** Define the log level for the in-cluster controllers */
+  logLevel: string;
+  /** Propagate env variables to in-cluster controllers */
+  env: Record<string, string>;
+  /** Custom RBAC rules */
+  rbac: PolicyRule[];
+  /** The RBAC mode; if "scoped", generates scoped rules, otherwise uses wildcard rules. */
+  rbacMode: string;
+  /** Custom Labels for Kubernetes Objects */
+  customLabels: CustomLabels;
 };
+/** Global configuration for the Pepr runtime. */
+export type ModuleConfig = {
+  /** A unique identifier for this Pepr module. This is automatically generated by Pepr. */
+  uuid: string;
+  /** Configure global exclusions that will never be processed by Pepr. */
+  alwaysIgnore: WebhookIgnore;
+} & Partial<ModuleConfigOptions>;
+
+export type PackageJSON = {
+  description: string;
+  pepr: ModuleConfig;
+};
+
+export type PeprModuleOptions = {
+  deferStart?: boolean;
+
+  /** A user-defined callback to pre-process or intercept a Pepr request from K8s immediately before it is processed */
+  beforeHook?: (req: AdmissionRequest) => void;
+
+  /** A user-defined callback to post-process or intercept a Pepr response just before it is returned to K8s */
+  afterHook?: (res: MutateResponse | ValidateResponse) => void;
+}; // Track if this is a watch mode controller
